requirements.txt
streamlit
pytesseract
pillow
openai
requests
pandas
